requirements.txt
streamlit
scikit-learn
joblib
matplotlib
wordcloud
PyPDF2
python-docx
pypiwin32
